mentorsData: Automatically generate an ID for each mentor
This ID is guaranteed to correspond to the mentor's actual index in the
`mentors` array.

This change fixes a bug where sorting the array led to discrepancies
between the specified IDs in the mentor objects and the actual indexes
of these objects in the `mentors` array, causing the wrong modal to be
opened.