Shift mentorship logo into `public` folder
The logo will be used in our meta tags, so we need it to be in the
`public` folder to allow the static HTML file to access the image. This
excludes the image from optimization and creating cached URLs, but since
this image will not change much (at most 1 per cycle), it does not
matter much.

Ref. https://create-react-app.dev/docs/using-the-public-folder