search-bar: Fine-tune lunr parameters
See https://lunrjs.com/guides/customising.html for the technical details
of the b and k1 parameters. In short, reducing b to 0 means documents of
different lengths will not affect the importance of the term, and
reducing k1 to 0 will cause repeated words to saturate immediately, so
that they have no effect on the term.